Kernel/univ: allow algebraic universes everywhere
This changes the interface of universes so that
universe polymorphic definitions can take arbitrary
universes as arguments, not just levels. This will
allow less constraints to be generated in many cases
where max() appears.

The graph implementation is changed so that l+n nodes
are accepted, with a specific semantics:
n < n+1 < .. < n +k is maintained as soon as n+k appears
in the graph. The implementation currently does not ensure
that no universe is between n and n+1 though.