Equality: refine test for patterns, configuration of indexes
Take into account the transparent state of "rewrite" hint db for
indexing and unification. Now by default allowing to unfold local
variables during unification. Rrespect the complete application
structure of the inferred pat. This allows to rewrite with a lemma
like (lem (_ _)) ensuring we find occurrences of the explicit
application using first-order unification as long as it is under an
indexed part of the lemma.