Implementation of induction-recursion and induction-induction
Rebased on trunk

Code refactoring around the is_positivity_check flag

Fix type-checking and validation of ind-rec definitions.

Fixes to go up to parsing of fixpoint definitions.

Complicated by the named/rel/global differences.

Typechecking first example, except for final retypechecking.

Main issues are:
- Global env needs to be modified to parse function pattern-matching
(globalization of constructor names)
- Switching btween de Bruijn/named/globalized entries.
- Allowing to refine a param to a term would maybe help ?

Typechecking the first inductive-recursive definition

This takes care of typechecking, guard included, but is rather
hackish in the way we treat the environment etc...
In the end, both the inductive and the recursive definition have
global names, so the recursion is tied at the level of the environment
(two global declarations refer to each other).

The positivity check is _not_ done.

First inductive-inductive definition.

This is clearly less intrusive than ind-rec.
We only allow dependencies between arities and
constructors of distinct inductive types (in text order).

Kernel: change representation of fixpoints

Now the type of later fixpoints can depend on the
type of previous ones.

Internalization/detyping: fixpoint representation

Kernel: new representation of fixpoints

push_rec_types/push_corec_types distinction,
more code needs to be adapted, added FIXMEs there

Ind-ind: fixpoints depending on each other

Pretyping is done differently, to tie the knot of recursion.
- First we typecheck the types of the block of fixpoints,
  later ones can depend on previous ones.
- Then we build fixpoint block with these types and evars
  for the bodies.
- We build a rel_context containing the projections of that
  block.
- We typecheck each definition in turn, instantiating the
  corresponding evar. This allows later definitions to depend
  on the computational behavior of previous ones (needed for
  the elimination principle).
- At the end we have a whole instantiated block to be sent to
  the kernel.
- In typeops, we currently have to check the fixpoint bodies
  assuming they are well-typed and well-founded. We could
  refine this to ensure that each block f_1..f_i is well-typed
  and guarded before typechecking f_i+1.
- Due to the new representation, when projecting out a particular
  fixpoint type, it must get substituted by the previous fixpoints
  it might depend on.

Fix typing with recursive definitions

Allow some unification to happen when comparing the
infered and ascribed types of fixpoints.

indrec

Fix representation: fix detyping

Fixes in typechecking of rec-rec definitions.

wrong lifting and substitutions.

Ind-ind: extend to internal constructor deps

This requires typechecking subsequent constructors
in extended environments, and gradually substitute
them in the kernel (not in one block as before).

- Had to use fold_left_map in the kernel for proper
construction of the substitution of de Bruijn indices
corresponding to constructors.

- To typecheck the associated elimination principle,
it is required to use unification constraints, as the
well-typedness of the body of a branch of the match
can depend directly on the content of previous ones.
When the whole fixpoint body is instantiated this
constraint can be solved. This goes a little beyond
what was necessary for using constructors of previous
inductives, where only the content of previous fixpoints
was needed.

Note that checking positivity and guardedness does not
seem more problematic than before. Same problem with the
return predicate of matches, in which we have recursive
calls now, the kernel is still unsafe in this respect
(we do not check predicates of cases at all right now).

Temporary fix for discharge, not working with ind-ind

Kernel: handle inductives with parameters

Instantiate constructors with params in other constructors.

Fix termops operations w.r.t. Fix types

Fix elaboration of rec-rec definitions.

Fix de Bruijn bug in inductive-recursive fixpoints

de Bruijn fixes

Adapt typeops to rec-rec definitions.

Better handling of the fixpoint list for ind-rec defs

Use {struct n} annotations when available when

typechecking internal fixes.