Evd/evarsolve: add a field to evars for unification
To know which variables come from abstractions (define_evar_as_lambda)
and should hence _not_ be subject to imitation when doing
inversion/second_order_matching but always abstracted.
This respects the unification of evars of arrow type better: one
expects a function whereas before an argument in the unification
problem could get imitated in the body.