Enhance [merge] for Stores used in [Evd.restrict] and evar-evar assignments
The only solution I found to preserve the easy extensibility is to
store the merge function alongside the value in the untyped store,
which results in duplication. Other ideas are welcome.
Now restricted evars inherit the store of their parent, and when
solving evar-evar problems the undefined evars gets a [merge]d store.